fix(otel): address staging pr comments for trigger otel (#4586)
